+++ src/pwanstrings.h
@@ -8,7 +8,7 @@ namespace pwan
 {
     namespace strings
     {
-        std::string                     fromInt(long long int number, unsigned int padding = 0, unsigned int base = 10);
+        std::string                     fromInt(int64_t number, unsigned int padding = 0, unsigned int base = 10);
         std::vector<std::string>        explode(const std::string& inputstring, std::string explodeby = " ");
         std::string                     toLower(std::string inputstring);
         std::string                     base64Encode(const std::string &text);
